+++ src/main/java/wf3/project/alpha_betise/repositories/PhotoLivreRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import wf3.project.alpha_betise.entities.PhotoLivre;
 
 @Repository
-public interface PhotoLivreRepository extends JpaRepository<PhotoLivre, Integer> {
+public interface PhotoLivreRepository extends JpaRepository<PhotoLivre, Long> {
 
 }
